+++ src/components/BackButton/BackButton.scss
@@ -1,0 +1,27 @@
+.Back-button {
+  &__back-arrow {
+    position: relative;
+  }
+
+  &__back-link {
+    display: inline-block;
+
+    font-family: 'Mont-SemiBold';
+    color: #89939A;
+    font-size: 12px;
+    font-weight: 700;
+    text-decoration: none;
+    background-color: #ffffff;
+    border: none;
+    cursor: pointer;
+
+    &::after {
+      position: absolute;
+      content: 'Back';
+    }
+
+    &:hover {
+      color: #0F0F11;
+    }
+  }
+}
