+++ src/components/ItemCard/EmptyCart/EmptyCart.scss
@@ -1,0 +1,12 @@
+.Empty-cart {
+  &__header {
+    color: #0F0F11;
+    text-align: center;
+
+    font-family: 'Mont-Bold';
+    font-size: 32px;
+    font-weight: 800;
+    line-height: 41px;
+    letter-spacing: -0.32px;
+  }
+}
