+++ src/components/ItemCard/EmptyCart/index.ts
@@ -1,0 +1,1 @@
+export * from './EmptyCart';
